requirements/tts.txt
fastapi==0.111.0
uvicorn[standard]==0.25.0
redis==5.0.1
numpy==1.24.0
soundfile==0.12.1
python-multipart==0.0.20
loguru==0.7.0
piper-tts
requests==2.31.0